Fixing a serious bug in the Follow command
The issue here is that the relations array was not getting properly extended.  The
call to copy wasn't doing what was expected.  I've replaced that code with something
that is probably less efficient, but easier to read and definitely works.  It
basically builds a whole new array by appending the previous elements and then
one additional element.

I also created some tests to make sure that a double follow works.  This involved
a couple of changes to the _links elements in the test case.  It didn't change any
of the existing tests, just added a few new ones that leveraged the new link
relations that I added.